move remote settings API to remote component
