+++ lib/datadog/ci/configuration/components.rb
@@ -5,6 +5,7 @@ require "datadog/core/telemetry/ext"
 require_relative "../ext/settings"
 require_relative "../git/tree_uploader"
 require_relative "../remote/component"
+require_relative "../remote/library_settings_client"
 require_relative "../test_optimisation/component"
 require_relative "../test_optimisation/coverage/transport"
 require_relative "../test_optimisation/coverage/writer"
@@ -16,7 +17,6 @@ require_relative "../test_visibility/serializers/factories/test_suite_level"
 require_relative "../test_visibility/transport"
 require_relative "../transport/adapters/telemetry_webmock_safe_adapter"
 require_relative "../transport/api/builder"
-require_relative "../transport/remote_settings_api"
 require_relative "../utils/identity"
 require_relative "../utils/parsing"
 require_relative "../utils/test_run"
@@ -108,7 +108,7 @@ module Datadog
 
           @git_tree_upload_worker = build_git_upload_worker(settings, test_visibility_api)
           @ci_remote = Remote::Component.new(
-            library_settings_api: build_remote_settings_client(settings, test_visibility_api)
+            library_settings_client: build_library_settings_client(settings, test_visibility_api)
           )
           # @type ivar @test_optimisation: Datadog::CI::TestOptimisation::Component
           @test_optimisation = build_test_optimisation(settings, test_visibility_api)
@@ -220,8 +220,8 @@ module Datadog
           end
         end
 
-        def build_remote_settings_client(settings, api)
-          Transport::RemoteSettingsApi.new(
+        def build_library_settings_client(settings, api)
+          Remote::LibrarySettingsClient.new(
             api: api,
             dd_env: settings.env,
             config_tags: custom_configuration(settings)
