+++ lib/datadog/ci/remote/component.rb
@@ -6,13 +6,13 @@ module Datadog
       # Remote configuration component.
       # Responsible for fetching library settings and configuring the library accordingly.
       class Component
-        def initialize(library_settings_api:)
-          @library_settings_api = library_settings_api
+        def initialize(library_settings_client:)
+          @library_settings_client = library_settings_client
         end
 
         # called on test session start, uses test session info to send configuration request to the backend
         def configure(test_session)
-          library_configuration = @library_settings_api.fetch_library_settings(test_session)
+          library_configuration = @library_settings_client.fetch(test_session)
           # sometimes we can skip code coverage for default branch if there are no changes in the repository
           # backend needs git metadata uploaded for this test session to check if we can skip code coverage
           if library_configuration.require_git?
@@ -20,7 +20,7 @@ module Datadog
             git_tree_upload_worker.wait_until_done
 
             Datadog.logger.debug { "Requesting library configuration again..." }
-            library_configuration = @library_settings_api.fetch_library_settings(test_session)
+            library_configuration = @library_settings_client.fetch(test_session)
 
             if library_configuration.require_git?
               Datadog.logger.debug { "git metadata upload did not complete in time when configuring library" }
