+++ sig/datadog/ci/configuration/components.rbs
@@ -31,7 +31,7 @@ module Datadog
 
         def build_git_upload_worker: (untyped settings, Datadog::CI::Transport::Api::Base? api) -> Datadog::CI::Worker
 
-        def build_remote_settings_client: (untyped settings, Datadog::CI::Transport::Api::Base? api) -> Datadog::CI::Transport::RemoteSettingsApi
+        def build_library_settings_client: (untyped settings, Datadog::CI::Transport::Api::Base? api) -> Datadog::CI::Remote::LibrarySettingsClient
 
         def custom_configuration: (untyped settings) -> Hash[String, String]
