+++ sig/datadog/ci/remote/library_settings.rbs
@@ -1,0 +1,35 @@
+module Datadog
+  module CI
+    module Remote
+      class LibrarySettings
+        @http_response: Datadog::CI::Transport::Adapters::Net::Response?
+        @json: Hash[String, untyped]?
+
+        @require_git: bool
+        @itr_enabled: bool
+        @code_coverage_enabled: bool
+        @tests_skipping_enabled: bool
+
+        def initialize: (Datadog::CI::Transport::Adapters::Net::Response? http_response) -> void
+
+        def ok?: () -> bool
+
+        def payload: () -> Hash[String, untyped]
+
+        def require_git?: () -> bool
+
+        def itr_enabled?: () -> bool
+
+        def code_coverage_enabled?: () -> bool
+
+        def tests_skipping_enabled?: () -> bool
+
+        private
+
+        def bool: (String key) -> bool
+
+        def default_payload: () -> Hash[String, untyped]
+      end
+    end
+  end
+end
