--- sig/datadog/ci/transport/remote_settings_api.rbs
@@ -1,49 +1,0 @@
-module Datadog
-  module CI
-    module Transport
-      class RemoteSettingsApi
-        class Response
-          @http_response: Datadog::CI::Transport::Adapters::Net::Response?
-          @json: Hash[String, untyped]?
-
-          @require_git: bool
-          @itr_enabled: bool
-          @code_coverage_enabled: bool
-          @tests_skipping_enabled: bool
-
-          def initialize: (Datadog::CI::Transport::Adapters::Net::Response? http_response) -> void
-
-          def ok?: () -> bool
-
-          def payload: () -> Hash[String, untyped]
-
-          def require_git?: () -> bool
-
-          def itr_enabled?: () -> bool
-
-          def code_coverage_enabled?: () -> bool
-
-          def tests_skipping_enabled?: () -> bool
-
-          private
-
-          def bool: (String key) -> bool
-
-          def default_payload: () -> Hash[String, untyped]
-        end
-
-        @api: Datadog::CI::Transport::Api::Base?
-        @dd_env: String?
-        @config_tags: Hash[String, String]
-
-        def initialize: (?api: Datadog::CI::Transport::Api::Base?, dd_env: String?, ?config_tags: Hash[String, String]?) -> void
-
-        def fetch_library_settings: (Datadog::CI::TestSession test_session) -> Response
-
-        private
-
-        def payload: (Datadog::CI::TestSession test_session) -> String
-      end
-    end
-  end
-end
