+++ spec/datadog/ci/remote/component_spec.rb
@@ -3,9 +3,9 @@
 require_relative "../../../../lib/datadog/ci/remote/component"
 
 RSpec.describe Datadog::CI::Remote::Component do
-  subject(:component) { described_class.new(library_settings_api: library_settings_api) }
+  subject(:component) { described_class.new(library_settings_client: library_settings_client) }
 
-  let(:library_settings_api) { instance_double(Datadog::CI::Transport::RemoteSettingsApi) }
+  let(:library_settings_client) { instance_double(Datadog::CI::Remote::LibrarySettingsClient) }
   let(:git_tree_upload_worker) { instance_double(Datadog::CI::Worker) }
   let(:test_optimisation) { instance_double(Datadog::CI::TestOptimisation::Component) }
 
@@ -19,11 +19,11 @@ RSpec.describe Datadog::CI::Remote::Component do
 
     let(:test_session) { instance_double(Datadog::CI::TestSession) }
     let(:library_configuration) do
-      instance_double(Datadog::CI::Transport::RemoteSettingsApi::Response, require_git?: require_git)
+      instance_double(Datadog::CI::Remote::LibrarySettings, require_git?: require_git)
     end
 
     before do
-      expect(library_settings_api).to receive(:fetch_library_settings)
+      expect(library_settings_client).to receive(:fetch)
         .with(test_session).and_return(library_configuration).once
     end
 
@@ -42,7 +42,7 @@ RSpec.describe Datadog::CI::Remote::Component do
 
       before do
         expect(git_tree_upload_worker).to receive(:wait_until_done)
-        expect(library_settings_api).to receive(:fetch_library_settings)
+        expect(library_settings_client).to receive(:fetch)
           .with(test_session).and_return(library_configuration)
 
         expect(test_optimisation).to receive(:configure).with(library_configuration, test_session)
